Delete new batch.js file created and move code from bloomstack_core batch.js file code in erpnext and move funtion into utils.py file
