+++ erpnext/utils.py
@@ -1,6 +1,7 @@
 import json
 import frappe
 from frappe.utils import get_url
+from erpnext.stock.doctype.batch.batch import get_batch_qty
 
 @frappe.whitelist()
 def get_contact(doctype, name, contact_field):
@@ -53,3 +54,29 @@ def create_authorization_request(dt, dn, contact_email, contact_name):
 	new_authorization_request.authorizer_email = contact_email
 	new_authorization_request.authorizer_name = contact_name
 	new_authorization_request.save()
+
+
+@frappe.whitelist()
+def move_expired_batches(source_name, target_doc=None):
+	batch_details = get_batch_qty(source_name)
+	target_warehouse = frappe.flags.args.get("warehouse")
+
+	item = frappe.db.get_value("Batch", source_name, "item")
+	uom = frappe.db.get_value("Item", item, "stock_uom")
+
+	stock_entry = frappe.new_doc("Stock Entry")
+	stock_entry.stock_entry_type = "Material Transfer"
+
+	for batch in batch_details:
+		if batch.get("qty") > 0:
+			stock_entry.append("items", {
+				"item_code": item,
+				"qty": batch.get("qty"),
+				"uom": uom,
+				"stock_uom": uom,
+				"batch_no": source_name,
+				"s_warehouse": batch.get("warehouse"),
+				"t_warehouse": target_warehouse
+			})
+
+	return stock_entry
